feat: set ID and description for VIP 483 and update the receiver wallet
