+++ simulations/vip-483/bscmainnet.ts
@@ -5,44 +5,36 @@ import { ethers } from "hardhat";
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import {
-  BSCMAINNET_USDT,
-  RISK_FUND,
-  USDT_TOKENS_AMOUNT,
-  VANGUARD_TREASURY,
-  vip474,
-} from "../../vips/vip-474/bscmainnet";
+import { BSCMAINNET_USDT, DEV_WALLET, RISK_FUND, USDT_TOKENS_AMOUNT, vip483 } from "../../vips/vip-483/bscmainnet";
 import RISK_FUND_ABI from "./abi/riskFund.json";
 import USDT_ABI from "./abi/usdt.json";
 
-forking(48003323, async () => {
+forking(48473745, async () => {
   const provider = ethers.provider;
   let usdt: Contract;
-  let treasuryBalanceBefore: BigNumber;
-  let treasuryBalanceAfter: BigNumber;
+  let devWalletBalanceBefore: BigNumber;
   let riskFundBalanceBefore: BigNumber;
-  let riskFundBalanceAfter: BigNumber;
 
   before(async () => {
     usdt = new ethers.Contract(BSCMAINNET_USDT, USDT_ABI, provider);
-    treasuryBalanceBefore = await usdt.balanceOf(VANGUARD_TREASURY);
+    devWalletBalanceBefore = await usdt.balanceOf(DEV_WALLET);
     riskFundBalanceBefore = await usdt.balanceOf(RISK_FUND);
   });
 
-  testVip("VIP-474", await vip474(), {
+  testVip("VIP-483", await vip483(), {
     callbackAfterExecution: async (txResponse: TransactionResponse) => {
       await expectEvents(txResponse, [RISK_FUND_ABI], ["SweepTokenFromPool"], [1]);
     },
   });
 
   describe("Post-VIP behavior", () => {
-    it("balance of treasury after sweeping tokens", async () => {
-      treasuryBalanceAfter = await usdt.balanceOf(VANGUARD_TREASURY);
-      expect(treasuryBalanceAfter).equals(treasuryBalanceBefore.add(USDT_TOKENS_AMOUNT));
+    it("balance of devWallet after sweeping tokens", async () => {
+      const devWalletBalanceAfter = await usdt.balanceOf(DEV_WALLET);
+      expect(devWalletBalanceAfter).equals(devWalletBalanceBefore.add(USDT_TOKENS_AMOUNT));
     });
 
     it("balance of risk fund after sweeping tokens", async () => {
-      riskFundBalanceAfter = await usdt.balanceOf(RISK_FUND);
+      const riskFundBalanceAfter = await usdt.balanceOf(RISK_FUND);
       expect(riskFundBalanceAfter).equals(riskFundBalanceBefore.sub(USDT_TOKENS_AMOUNT));
     });
   });
